+++ lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_exceptions.rb
@@ -3,4 +3,6 @@
 module OnlyofficeWebdriverWrapper
   # Exception class if system is not supported
   class WebdriverSystemNotSupported < StandardError; end
+  # If entry needed to be selected not found in select
+  class SelectEntryNotFound < StandardError; end
 end